remove use of feeluown.models
